# morphio-io/frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Morphio.io API",
		"description": "This API will blow your mind",
		"version": "1.0.0"
	},
	"paths": {
		"/auth/login": {
			"post": {
				"tags": ["auth", "Authentication"],
				"summary": "Login",
				"operationId": "login",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLogin"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Login",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthTokenResponse"
								},
								"example": {
									"access_token": "xxx",
									"user": {
										"id": 1,
										"email": "user1@example.com",
										"display_name": "user1"
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"tags": ["auth", "Authentication"],
				"summary": "Logout",
				"operationId": "logout",
				"responses": {
					"200": {
						"description": "Successfully Logged Out",
						"content": {
							"application/json": {
								"schema": {},
								"example": {
									"status": "success",
									"message": "Logged out successfully."
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
...
		"/auth/refresh-token": {
			"post": {
				"tags": ["auth", "Authentication"],
				"summary": "Refresh Token",
				"operationId": "refresh_token",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthTokenResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					}
				}
			}
		},
...
		"/admin/get-usage": {
			"get": {
				"tags": ["Admin", "Admin"],
				"summary": "Get Usage",
				"description": "Return a list of usage stats (per user + usage_type).\nIncludes the user's email & display name for better tracking.",
				"operationId": "get_admin_usage",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_list_dict__"
								}
							}
						}
					},
					"403": {
						"description": "Not authorized"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/admin/get-subscriptions": {
			"get": {
				"tags": ["Admin", "Admin"],
				"summary": "Get Subscriptions",
				"operationId": "get_subscriptions",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_list_SubscriptionOut__"
								}
							}
						}
					},
					"403": {
						"description": "Not authorized"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
			"AuthTokenPayload": {
				"properties": {
					"access_token": {
						"type": "string",
						"title": "Access Token"
					},
					"refresh_token": {
						"type": "string",
						"title": "Refresh Token",
						"default": ""
					},
					"user": {
						"$ref": "#/components/schemas/app__schemas__auth_schema__UserOut"
					}
				},
				"type": "object",
				"required": ["access_token", "user"],
				"title": "AuthTokenPayload",
				"description": "Payload for auth token responses (login, register, refresh)."
			},
			"AuthTokenResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ResponseStatus"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/AuthTokenPayload"
							},
							{
								"type": "null"
							}
						]
					},
					"timestamp": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Timestamp"
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "AuthTokenResponse",
				"description": "Wrapped response for auth token endpoints."
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"ResponseModel_list_SubscriptionOut__": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/SubscriptionOut"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Data"
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[list[SubscriptionOut]]"
			},
			"ResponseModel_list_dict__": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"additionalProperties": true,
									"type": "object"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Data"
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[list[dict]]"
			},
			"ResponseStatus": {
				"type": "string",
				"enum": ["success", "error", "warning", "info"],
				"title": "ResponseStatus"
			},
			"SubscriptionOut": {
				"properties": {
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"user_id": {
						"type": "integer",
						"title": "User Id"
					},
					"plan": {
						"type": "string",
						"title": "Plan"
					},
					"status": {
						"type": "string",
						"title": "Status"
					},
					"stripe_customer_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Stripe Customer Id"
					},
					"stripe_subscription_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Stripe Subscription Id"
					},
					"created_at": {
						"type": "string",
						"title": "Created At"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Updated At"
					},
					"user": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/UserBasicOut"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": ["id", "user_id", "plan", "status", "created_at"],
				"title": "SubscriptionOut"
			},
...
			"UserBasicOut": {
				"properties": {
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"email": {
						"type": "string",
						"title": "Email"
					},
					"display_name": {
						"type": "string",
						"title": "Display Name"
					}
				},
				"type": "object",
				"required": ["id", "email", "display_name"],
				"title": "UserBasicOut"
			},
...
			"UserLogin": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"title": "Email"
					},
					"password": {
						"type": "string",
						"title": "Password"
					}
				},
				"type": "object",
				"required": ["email", "password"],
				"title": "UserLogin"
			},
			"UserRole": {
				"type": "string",
				"enum": ["USER", "ADMIN", "POWER", "MODERATOR", "GUEST"],
				"title": "UserRole"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
			"app__schemas__auth_schema__UserOut": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"title": "Email"
					},
					"display_name": {
						"type": "string",
						"maxLength": 50,
						"minLength": 1,
						"title": "Display Name"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole",
						"default": "USER"
					},
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"title": "Created At"
					},
					"last_login": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Last Login"
					},
					"is_active": {
						"type": "boolean",
						"title": "Is Active"
					},
					"content_count": {
						"type": "integer",
						"title": "Content Count",
						"default": 0
					}
				},
				"type": "object",
				"required": ["email", "display_name", "id", "created_at", "is_active"],
				"title": "UserOut"
			}
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": {
					"password": {
						"scopes": {},
						"tokenUrl": "/auth/login"
					}
				}
			}
		}
	}
}